[!] Fifth prototype version.
[+] GetBoardConfiguration().
[^] Code structure optimizations.

It's not release-ready yet.

Mininum set of things to do:
  [ ] Use GetBoardConfiguration() in logic.
  [ ] Cleanup code from t2s().
  [ ] Move Complain() to [workshop].

-- Some selected messages as a reminder --

[^-] Sysex data field removed..

Sysex commands framing and wrapping is responsibility of [Transmitter]
now.

Separating Sysex commands is not needed as long as that commands
lie in data range (0x00..0x7F) and so do not intersect with
ordinary commands (range 0x80..0xFF).

[-] [CLI_Blink] General functionality covered in [CLI_SmokeTest].
[+] [CreateDeploy] Added side modules for DS3231.
[~] [workshop] Lotta new files due t2s() and DS3231 codec.
[!/] Complain() function made global..

Temporarily defined in [FirmataAmbassador.Interface].